refactor(l1): cleanup storage crate. (#2049)
**Motivation**
Prepare the ground to divide the store into multiple stores (state,
history, etc)

**Description**
- Created `lib.rs` file that exposed the modules to the outside world.
- Renamed `engines` to `store_db` to be more aligned with `trie_db`
- Renamed `storage.rs` to `store.rs`
