+++ crates/storage/store/Cargo.toml
@@ -33,4 +33,4 @@ hex-literal.workspace = true
 tempdir = "0.3.7"
 
 [lib]
-path = "./storage.rs"
+path = "./lib.rs"
