+++ crates/storage/store/store_db/in_memory.rs
@@ -1,4 +1,8 @@
-use crate::{error::StoreError, MAX_SNAPSHOT_READS, STATE_TRIE_SEGMENTS};
+use crate::{
+    api::StoreEngine,
+    error::StoreError,
+    store::{MAX_SNAPSHOT_READS, STATE_TRIE_SEGMENTS},
+};
 use bytes::Bytes;
 use ethereum_types::{H256, U256};
 use ethrex_common::types::{
@@ -11,8 +15,6 @@ use std::{
     fmt::Debug,
     sync::{Arc, Mutex, MutexGuard},
 };
-
-use super::api::StoreEngine;
 
 pub type NodeMap = Arc<Mutex<HashMap<Vec<u8>, Vec<u8>>>>;
