+++ crates/storage/store/store_db/mod.rs
@@ -1,7 +1,5 @@
-pub mod api;
 pub mod in_memory;
 #[cfg(feature = "libmdbx")]
 pub mod libmdbx;
 #[cfg(feature = "redb")]
 pub mod redb;
-mod utils;
